Feature/authorization

+++ api/app/Http/Controllers/EntityController.php
@@ -10,6 +10,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Extensions\Controller\AuthorizesRequestsTrait;
 use Illuminate\Http\Request;
 use Spira\Model\Model\BaseModel;
 use Elasticquent\ElasticquentTrait;
@@ -23,7 +24,7 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 abstract class EntityController extends ApiController
 {
-    use RequestValidationTrait;
+    use RequestValidationTrait, AuthorizesRequestsTrait;
 
     /**
      * @var BaseModel
@@ -46,6 +47,7 @@ abstract class EntityController extends ApiController
     {
         $collection = $this->getAllEntities();
         $collection = $this->getWithNested($collection, $request);
+        $this->authorize($collection);
 
         return $this->getResponse()
             ->transformer($this->getTransformer())
@@ -65,6 +67,7 @@ abstract class EntityController extends ApiController
         }
 
         $collection = $this->getWithNested($collection, $request);
+        $this->authorize($collection);
 
         return $this->getResponse()
             ->transformer($this->getTransformer())
@@ -82,6 +85,7 @@ abstract class EntityController extends ApiController
     {
         $model = $this->findOrFailEntity($id);
         $model = $this->getWithNested($model, $request);
+        $this->authorize($model);
 
         return $this->getResponse()
             ->transformer($this->getTransformer())
@@ -99,6 +103,7 @@ abstract class EntityController extends ApiController
         $model = $this->getModel()->newInstance();
         $this->validateRequest($request->json()->all(), $this->getValidationRules());
         $model->fill($request->json()->all());
+        $this->authorize($model);
         $model->save();
 
         return $this->getResponse()
@@ -116,12 +121,12 @@ abstract class EntityController extends ApiController
     public function putOne(Request $request, $id)
     {
         $this->checkEntityIdMatchesRoute($request, $id, $this->getModel());
-
         $model = $this->findOrNewEntity($id);
 
         $this->validateRequest($request->json()->all(), $this->getValidationRules());
 
         $model->fill($request->json()->all());
+        $this->authorize($model);
         $model->save();
 
         return $this->getResponse()
@@ -143,8 +148,11 @@ abstract class EntityController extends ApiController
         $existingModels = $this->findCollection($requestCollection);
 
         $modelCollection = $this->getModel()
-            ->hydrateRequestCollection($requestCollection, $existingModels)
-            ->each(function (BaseModel $model) {
+            ->hydrateRequestCollection($requestCollection, $existingModels);
+
+        $this->authorize($modelCollection);
+
+        $modelCollection->each(function (BaseModel $model) {
                 return $model->save();
             });
 
@@ -169,6 +177,7 @@ abstract class EntityController extends ApiController
         $this->validateRequest($request->json()->all(), $this->getValidationRules(), true);
 
         $model->fill($request->json()->all());
+        $this->authorize($model);
         $model->save();
 
         return $this->getResponse()->noContent();
@@ -188,9 +197,12 @@ abstract class EntityController extends ApiController
 
         $existingModels = $this->findOrFailCollection($requestCollection);
 
-        $this->getModel()
-            ->hydrateRequestCollection($requestCollection, $existingModels)
-            ->each(function (BaseModel $model) {
+        $modelsCollection = $this->getModel()
+            ->hydrateRequestCollection($requestCollection, $existingModels);
+
+        $this->authorize($existingModels);
+
+        $modelsCollection->each(function (BaseModel $model) {
                 return $model->save();
             });
 
@@ -205,7 +217,11 @@ abstract class EntityController extends ApiController
      */
     public function deleteOne($id)
     {
-        $this->findOrFailEntity($id)->delete();
+        $entity = $this->findOrFailEntity($id);
+
+        $this->authorize($entity);
+
+        $entity->delete();
 
         return $this->getResponse()->noContent();
     }
@@ -220,8 +236,11 @@ abstract class EntityController extends ApiController
     {
         $requestCollection = $request->json()->all();
 
-        $this->findOrFailCollection($requestCollection)
-            ->each(function (BaseModel $model) {
+        $modelsCollection = $this->findOrFailCollection($requestCollection);
+
+        $this->authorize($modelsCollection);
+
+        $modelsCollection->each(function (BaseModel $model) {
                 $model->delete();
             });
